DPproxy/requirements.txt
Django==1.11.4
django-cors-headers==2.1.0
django-filter==1.0.4
django-rest-auth==0.9.1
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
PyJWT==1.5.2
pytz==2017.2
six==1.10.0